+++ tests/testthat/test-feature-n-obs.R
@@ -1,5 +1,5 @@
 wages_test <- sample_frac_keys(wages, 0.05)
-df_n_obs <- features(wages_test, id, n_obs)
+df_n_obs <- features(wages_test, ln_wages, n_obs)
 new_dims <- c(n_keys(wages_test), 2)
 
 test_that("feature returns the right dimensions", {
@@ -15,8 +15,6 @@ test_that("longnostic returns a tbl_df", {
 })
 
 test_that("longnostic returns correct classes", {
-  expect_equal(classes(df_l_diff_1), 
-               c(id = "integer", V1 = "numeric"))
   expect_equal(classes(df_n_obs),
                c(id = "integer", n_obs = "integer"))
 })