+++ tests/testthat/test-if_tsibble.R
@@ -1,3 +1,5 @@
+context("if_tsibble")
+
 test_that("test_if_tsibble works", {
   expect_silent(test_if_tsibble(wages))
   expect_error(test_if_tsibble(iris))